Change Show.most_recent_source_created_at to most_recent_source_updated_at
This better reflects the semantics of what we want. This can be used to
determine the age of the source most recently published published at
upstream sources. `updated_at`
for a source comes from the upstream source and won't be suffer
recently imported biases.

The value pulled from upstream is based on when that source was
published/updated, not
counting reviews, ratings, etc.

How does this sound @switz? I hesitate to call it date published
because that isn’t always available or it is unreliable. This also lets
us surface shows that were fixed/improved after initial publish.